Moved some stuff to shared. Reorganized a bit. I could be here all day...
