+++ Content.Shared/_Floof/OfferItem/OfferableVirtualItemComponent.cs
@@ -8,6 +8,4 @@ namespace Content.Shared._Floof.OfferItem;
 /// </summary>
 [RegisterComponent]
 [NetworkedComponent]
-public sealed partial class OfferableVirtualItemComponent : Component
-{
-}
+public sealed partial class OfferableVirtualItemComponent : Component;
