+++ Content.Shared/_Floof/OfferItem/SharedOfferItemSystem.cs
@@ -1,7 +1,11 @@
+using Content.Shared._DV.Carrying;
+using Content.Shared.Alert;
 using Content.Shared.Hands.Components;
 using Content.Shared.Hands.EntitySystems;
 using Content.Shared.IdentityManagement;
 using Content.Shared.Interaction;
+using Content.Shared.Movement.Pulling.Components;
+using Content.Shared.Movement.Pulling.Systems;
 using Content.Shared.Nutrition.EntitySystems;
 using Robust.Shared.Network;
 using Robust.Shared.Player;
@@ -13,6 +17,7 @@ using Robust.Shared.Timing;
 // I did what I could to document the parts I managed to understand, but there is still more truth to be unveiled.
 //
 // HOURS_WASTED_HERE_FLOOFSTATION = 10
+// HOURS_WASTED_HERE_DELTAV = 1
 
 namespace Content.Shared._Floof.OfferItem;
 
@@ -22,14 +27,58 @@ public abstract partial class SharedOfferItemSystem : EntitySystem
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly SharedTransformSystem _transform = default!;
     [Dependency] private readonly SharedHandsSystem _hands = default!;
+    [Dependency] private readonly AlertsSystem _alertsSystem = default!;
+    [Dependency] private readonly CarryingSystem _carrying = default!;
+    [Dependency] private readonly PullingSystem _pulling = default!;
 
     public override void Initialize()
     {
+        base.Initialize();
+        SubscribeLocalEvent<OfferItemComponent, AcceptOfferAlertEvent>(OnAcceptOffer);
         SubscribeLocalEvent<OfferItemComponent, InteractUsingEvent>(OnInteractWithReceiver, before: [typeof(IngestionSystem)]);
         SubscribeLocalEvent<OfferableVirtualItemComponent, BeforeRangedInteractEvent>(OnRangedInteractWithReceiver);
         SubscribeLocalEvent<OfferItemComponent, MoveEvent>(OnMove);
 
+        SubscribeLocalEvent<BeingCarriedComponent, ItemTransferredEvent>(OnCarryTransfer);
+        SubscribeLocalEvent<PullableComponent, ItemTransferredEvent>(OnPulledTransfer);
+
         InitializeInteractions();
+    }
+
+    public override void Update(float frameTime)
+    {
+        base.Update(frameTime);
+
+        var query = EntityQueryEnumerator<OfferItemComponent, HandsComponent>();
+        while (query.MoveNext(out var uid, out var offerItem, out var hands))
+        {
+            // If the mob no longer holds an item in the original offering hand, clear offering mode
+            if (offerItem.Hand != null && !_hands.TryGetHeldItem((uid, hands), offerItem.Hand, out _))
+            {
+                if (offerItem.ReceivingFrom != null)
+                {
+                    UnReceive(offerItem.ReceivingFrom.Value, offererComp: offerItem);
+                    offerItem.IsInOfferMode = false;
+                    Dirty(uid, offerItem);
+                }
+                else
+                    UnOffer(uid, offerItem);
+            }
+
+            if (!offerItem.IsInReceiveMode)
+            {
+                _alertsSystem.ClearAlert(uid, offerItem.OfferAlert);
+                continue;
+            }
+
+            _alertsSystem.ShowAlert(uid, offerItem.OfferAlert);
+        }
+    }
+
+    #region Events
+    private void OnAcceptOffer(Entity<OfferItemComponent> ent, ref AcceptOfferAlertEvent args)
+    {
+        Receive((ent, ent.Comp));
     }
 
     private void OnInteractWithReceiver(Entity<OfferItemComponent> receiver, ref InteractUsingEvent args)
@@ -78,6 +127,46 @@ public abstract partial class SharedOfferItemSystem : EntitySystem
         args.Handled = CreateOffer((receiver.Value, receiverComponent), (offerer, offererComponent));
     }
 
+    private void OnMove(EntityUid uid, OfferItemComponent component, MoveEvent args)
+    {
+        if (_net.IsClient) // Client often mispredicts movement, we cant trust it here
+            return;
+
+        if (component.ReceivingFrom == null)
+            return;
+
+        if (_transform.InRange(args.NewPosition, Transform(component.ReceivingFrom.Value).Coordinates, component.MaxOfferDistance))
+            return;
+
+        UnOffer(uid, component);
+    }
+
+    private void OnCarryTransfer(Entity<BeingCarriedComponent> ent, ref ItemTransferredEvent args)
+    {
+        if (args.Handled
+            || args.PassedItem == args.RealItem // Means the entity is transferred NOT via carrying
+            || args.RealItem is not { Valid: true } carried
+            || ent.Comp.Carrier is not { Valid: true } oldCarrier)
+            return;
+
+        _carrying.DropCarried(oldCarrier, ent);
+        args.Handled = _carrying.TryCarry(args.Target, carried);
+    }
+
+    private void OnPulledTransfer(Entity<PullableComponent> ent, ref ItemTransferredEvent args)
+    {
+        if (args.Handled
+            || args.PassedItem == args.RealItem // Means the entity is transferred NOT via pulling
+            || args.RealItem is not { Valid: true } pulled)
+            return;
+
+        _pulling.TryStopPull(pulled, ent);
+        args.Handled = _pulling.TryStartPull(args.Target, ent, null, ent.Comp);
+    }
+
+    #endregion
+
+    #region Offering / Recieving
     /// <summary>
     ///     Attempts to create an offer. Expects offerer.Item to already be set to the offered item, offererComponent.InReceiveMode == true.
     ///     Will fail if offerer == receiver or if receiver already has a set TargetOrOfferer, and that person is not the current offerer
@@ -86,8 +175,10 @@ public abstract partial class SharedOfferItemSystem : EntitySystem
     {
         var offererComponent = offerer.Comp;
         var receiverComponent = receiver.Comp;
-        if (offerer == receiver || receiverComponent.IsInReceiveMode || !offererComponent.IsInOfferMode ||
-            (offererComponent.IsInReceiveMode && offererComponent.ReceivingFrom != receiver))
+        if (offerer == receiver || receiverComponent.IsInReceiveMode || !offererComponent.IsInOfferMode)
+            return false;
+
+        if (offererComponent.IsInReceiveMode && offererComponent.ReceivingFrom != receiver)
             return false;
 
         receiverComponent.IsInReceiveMode = true;
@@ -116,33 +207,23 @@ public abstract partial class SharedOfferItemSystem : EntitySystem
                 ("user", Identity.Entity(receiverComponent.ReceivingFrom.Value, EntityManager)),
                 ("item", Identity.Entity(offererComponent.GetRealEntity(EntityManager), EntityManager))),
             offerer,
-            receiver);
+            receiver,
+            Popups.PopupType.Medium);
 
         return true;
     }
 
-    private void OnMove(EntityUid uid, OfferItemComponent component, MoveEvent args)
-    {
-        if (_net.IsClient) // Client often mispredicts movement, we cant trust it here
-            return;
 
-        if (component.ReceivingFrom == null ||
-            args.NewPosition.InRange(EntityManager, _transform,
-                Transform(component.ReceivingFrom.Value).Coordinates, component.MaxOfferDistance))
-            return;
-
-        UnOffer(uid, component);
-    }
 
     /// <summary>
-    /// Resets the <see cref="_Floof.OfferItem.OfferItemComponent"/> of the user and the target
+    /// Resets the <see cref="OfferItemComponent"/> of the user and the target
     /// </summary>
     protected void UnOffer(EntityUid thisEntity, OfferItemComponent offererComp)
     {
         if (!TryComp<HandsComponent>(thisEntity, out var hands) || _hands.GetActiveHand((thisEntity, hands)) is null)
             return;
 
-        if (offererComp.ReceivingFrom is {} otherEntity && TryComp<OfferItemComponent>(otherEntity, out var otherOfferer))
+        if (offererComp.ReceivingFrom is { } otherEntity && TryComp<OfferItemComponent>(otherEntity, out var otherOfferer))
         {
             // So this tries to figure out which of these entities do what...
             // if A.OfferItemComponent.Item != null, then A is currently offering an item to A.OfferItemComponent.TargetOrOfferer
@@ -244,10 +325,90 @@ public abstract partial class SharedOfferItemSystem : EntitySystem
     }
 
     /// <summary>
-    /// Returns true if <see cref="_Floof.OfferItem.OfferItemComponent.IsInOfferMode"/> = true
+    /// Accepting the offer and receive item
     /// </summary>
-    protected bool IsInOfferMode(EntityUid? entity, OfferItemComponent? component = null)
+    public void Receive(Entity<OfferItemComponent?> receiver)
     {
-        return entity != null && Resolve(entity.Value, ref component, false) && component.IsInOfferMode;
+        if (!_timing.IsFirstTimePredicted)
+            return;
+
+        if (!Resolve(receiver, ref receiver.Comp))
+            return;
+
+        if (!TryComp<OfferItemComponent>(receiver.Comp.ReceivingFrom, out var offererComponent) ||
+            offererComponent.Hand == null ||
+            receiver.Comp.ReceivingFrom is not {} sender ||
+            !TryComp<HandsComponent>(receiver, out var hands))
+            return;
+
+        if (offererComponent.Item != null)
+        {
+            // Floof - check if there's something else handling it first
+            var realItem = offererComponent.GetRealEntity(EntityManager);
+            if (!TryHandleExtendedTransfer(sender, receiver, offererComponent.Item.Value, realItem)
+                && !_hands.TryPickup(receiver, offererComponent.Item.Value, handsComp: hands))
+            {
+                _popup.PopupEntity(Loc.GetString("offer-item-full-hand"), receiver, receiver);
+                return;
+            }
+
+            _popup.PopupEntity(
+                Loc.GetString("offer-item-give",
+                    ("item", Identity.Entity(realItem, EntityManager)), // FLoof - resolve virtual items
+                    ("target", Identity.Entity(receiver, EntityManager))),
+                sender,
+                sender);
+            _popup.PopupEntity(
+                Loc.GetString("offer-item-give-other",
+                    ("user", Identity.Entity(receiver.Comp.ReceivingFrom.Value, EntityManager)),
+                    ("item", Identity.Entity(realItem, EntityManager)), // FLoof - resolve virtual items
+                    ("target", Identity.Entity(receiver, EntityManager))),
+                sender,
+                Filter.PvsExcept(sender, entityManager: EntityManager),
+                true);
+        }
+
+        offererComponent.Item = null;
+        UnReceive(receiver, receiver.Comp, offererComponent);
     }
+    #endregion
+    /// <summary>
+    /// Returns true if <see cref="OfferItemComponent.IsInOfferMode"/> = true
+    /// </summary>
+    protected bool IsInOfferMode(Entity<OfferItemComponent?> ent)
+    {
+        return Resolve(ent, ref ent.Comp, false) && ent.Comp.IsInOfferMode;
+    }
+
+    private bool TryHandleExtendedTransfer(EntityUid user, EntityUid target, EntityUid offeredItem, EntityUid realItem)
+    {
+        var ev = new ItemTransferredEvent
+        {
+            User = user,
+            Target = target,
+            PassedItem = offeredItem,
+            RealItem = realItem,
+        };
+        RaiseLocalEvent(realItem, ref ev);
+        return ev.Handled;
+    }
+}
+
+/// <summary>
+///     Raised on the entity that was transferred via item offering.
+/// </summary>
+[ByRefEvent]
+public sealed class ItemTransferredEvent : HandledEntityEventArgs
+{
+    public EntityUid User;
+    public EntityUid Target;
+
+    /// <summary>
+    ///     The actual item being passed around. Can be a virtual item.
+    /// </summary>
+    public EntityUid PassedItem;
+    /// <summary>
+    ///     If <see cref="PassedItem"/> is a virtual item, this field contains the real item that was transferred.
+    /// </summary>
+    public EntityUid? RealItem;
 }
