+++ Resources/Locale/en-US/_Floof/interaction/offer-item-system.ftl
@@ -2,16 +2,16 @@ offer-item-empty-hand = You don't have anything in your hand to give!
 
 offer-item-full-hand = Your hand isn't free to receive the item.
 
-offer-item-try-give = You offer {THE($item)} to {$target}
-offer-item-try-give-target = {$user} offers you {THE($item)}
+offer-item-try-give = You offer {THE($item)} to {$target}.
+offer-item-try-give-target = {CAPITALIZE(THE($user))} offers you {THE($item)}.
 
-offer-item-give = You handed {THE($item)} to {$target}
-offer-item-give-other = {$user} handed {THE($item)} to {$target}
-offer-item-give-target = {$user} handed you {THE($item)}
+offer-item-give = You handed {THE($item)} to {$target}.
+offer-item-give-other = {CAPITALIZE(THE($user))} handed {THE($item)} to {$target}.
+offer-item-give-target = {CAPITALIZE(THE($user))} handed you {THE($item)}.
 
-offer-item-no-give = You stop offering {THE($item)} to {$target}
-offer-item-no-give-target = {$user} is no longer offering {THE($item)} to you
+offer-item-no-give = You stop offering {THE($item)} to {$target}.
+offer-item-no-give-target = {CAPITALIZE(THE($user))} is no longer offering {THE($item)} to you.
 
 
-alerts-offer-name = Accept offer
+alerts-offer-name = Accept Offer
 alerts-offer-desc = Click this alert to accept the item offered to you.
